+++ src/views/DestinationDetailView.vue
@@ -16,6 +16,7 @@ import WeatherClimateWidget from '../components/WeatherClimateWidget.vue';
 import UserTravelStories from '../components/UserTravelStories.vue';
 import AudioGuidePlayer from '../components/AudioGuidePlayer.vue';
 import TravelExpensesExporter from '../components/TravelExpensesExporter.vue';
+import DestinationComparator from '../components/DestinationComparator.vue';
 
 const route = useRoute();
 const travelStore = useTravelStore();
@@ -62,6 +63,7 @@ const relatedPosts = computed(() => {
         <ItineraryBuilder :destination-id="destination.id" class="margin-top-calc" />
         <UserTravelStories :destination-id="destination.id" class="margin-top-calc" />
         <TravelMapVisualizer :highlight-dest-id="destination.id" class="margin-top-calc" />
+        <DestinationComparator class="margin-top-calc" />
         <TravelExpensesExporter :destination-id="destination.id" class="margin-top-calc" />
         <PackingChecklist class="margin-top-calc" />
       </div>